Added printf statement for user to run eval "$(minikube docker-env" to setup docker to use docker daemon of minikube Also, cobra commands were failing after re-factoring.. corrected that
